Updates names: Response instead of Request

+++ src/main/java/com/agm/DietControlSystem/model/dto/ProductDTO.java
@@ -6,14 +6,14 @@ public class ProductDTO {
 	private String name;
 	private String category;
 	private String brand;
-	private Float calories;
-	private Float proteins;
+	private Integer calories;
+	private Integer proteins;
 	private Integer referenceQty;
 	private String measureUnit;
 	
 	public ProductDTO() {}
 
-	public ProductDTO(Long id, String name, String category, String brand, Float calories, Float proteins,
+	public ProductDTO(Long id, String name, String category, String brand, Integer calories, Integer proteins,
 			Integer referenceQty, String measureUnit) {
 		this.id = id;
 		this.name = name;
@@ -57,19 +57,19 @@ public class ProductDTO {
 		this.brand = brand;
 	}
 
-	public Float getCalories() {
+	public Integer getCalories() {
 		return calories;
 	}
 
-	public void setCalories(Float calories) {
+	public void setCalories(Integer calories) {
 		this.calories = calories;
 	}
 
-	public Float getProteins() {
+	public Integer getProteins() {
 		return proteins;
 	}
 
-	public void setProteins(Float proteins) {
+	public void setProteins(Integer proteins) {
 		this.proteins = proteins;
 	}
